исправил замечания и добавил экран подробностей транзакций

+++ app/src/main/java/com/example/androidplatform/domain/models/account/Account.kt
@@ -3,12 +3,12 @@ package com.example.androidplatform.domain.models.account
 import com.example.androidplatform.domain.models.clients.Client
 
 data class Account(
-    val id: Long,
+    val id: Int,
     val client: Client?,
     val createdDate: String,
-    val currency: Short,
+    val currency: Int,
     val number: String,
     val name: String,
-    val balance: Long,
+    val balance: Int,
     val state: String
 )
