+++ app/src/main/java/com/example/androidplatform/domain/models/account/Currency.kt
@@ -1,6 +1,6 @@
 package com.example.androidplatform.domain.models.account
 
-enum class Currency(val code: Short, val currencyName: String) {
+enum class Currency(val code: Int, val currencyName: String) {
     RUB(643, "Российский Рубль"),
     USD(840, "Доллар США"),
     EUR(978, "Евро"),
