+++ app/src/main/java/com/example/androidplatform/domain/models/history/Transaction.kt
@@ -1,14 +1,15 @@
 package com.example.androidplatform.domain.models.history
 
+import android.os.Parcelable
 import com.example.androidplatform.domain.models.account.Account
 
 data class Transaction (
-    val id: Long,
+    val id: Int,
     val account: Account,
     val receiver: String?,
     val date: String,
     val paymentDate: String,
-    val amount: Long,
+    val amount: Int,
     val comment: String,
     val reason: String?,
     val state: String,
